requirements.txt
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.136.1
gitdb==4.0.12
GitPython==3.1.49
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.4.4
openai==2.33.0
packaging==26.2
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
pypdf==6.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==1.0.0
streamlit==1.57.0
tenacity==9.1.4
toml==0.10.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
watchdog==6.0.0
websockets==16.0
